backend/requirements/dev.txt
-i https://pypi.org/simple/
Django==2.2.16
psycopg2==2.8 # в последнем обновлении етой библиотеки баг в тайм зонах
pytz
sqlparse
cryptography==2.8
Pillow
django-allauth
django-rest-auth[with_social]
dj-rest-auth  # token
django-jsonfield  #feald model
django-multiselectfield   #feald model
django-phonenumber-field  # feald model
phonenumbers  # feald model
djangorestframework
drf-yasg  # swagger
django-cors-headers
django-cookies-samesite
django-filter  # for params filter
liqpay-sdk-python3 # liqpay payment system
django-environ # for .env
google-cloud-storage # for google cloud

